PIM-3470 : move group and variant group denormalizer, re-work the abstract processor
