+++ spec/Pim/Bundle/BaseConnectorBundle/Processor/Denormalization/Flat/VariantGroupProcessorSpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace spec\Pim\Bundle\BaseConnectorBundle\Processor\ArrayToObject\Flat;
+namespace spec\Pim\Bundle\BaseConnectorBundle\Processor\Denormalization;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Bundle\BatchBundle\Item\InvalidItemException;
@@ -39,7 +39,8 @@ class VariantGroupProcessorSpec extends ObjectBehavior
             $valueNormalizer,
             $detacher,
             $groupClass,
-            $templateClass
+            $templateClass,
+            'csv'
         );
         $this->setStepExecution($stepExecution);
         $validator->validate(Argument::any())->willReturn(new ConstraintViolationList());
