+++ src/Pim/Bundle/BaseConnectorBundle/Processor/Denormalization/GroupProcessor.php
@@ -1,13 +1,14 @@
 <?php
 
-namespace Pim\Bundle\BaseConnectorBundle\Processor\ArrayToObject\Flat;
+namespace Pim\Bundle\BaseConnectorBundle\Processor\Denormalization;
 
-use Pim\Bundle\BaseConnectorBundle\Processor\ArrayToObject\AbstractProcessor;
+use Akeneo\Bundle\StorageUtilsBundle\Doctrine\ObjectDetacherInterface;
 use Pim\Bundle\CatalogBundle\Model\GroupInterface;
+use Pim\Bundle\CatalogBundle\Repository\ReferableEntityRepositoryInterface;
+use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
+use Symfony\Component\Validator\ValidatorInterface;
 
 /**
- * TODO : not sure about Flat folder cause we could "just" change the normalizer to use flat or structured
- *
  * Group import processor, allows to,
  *  - create / update groups (except variant group)
  *  - return the valid groups, throw exceptions to skip invalid ones
@@ -26,6 +27,29 @@ class GroupProcessor extends AbstractProcessor
 
     /** @staticvar string */
     const LABEL_FIELD = 'label';
+
+    /** @var string */
+    protected $format;
+
+    /**
+     * @param ReferableEntityRepositoryInterface $repository   repository to search the object in
+     * @param ValidatorInterface                 $validator    validator of the object
+     * @param DenormalizerInterface              $denormalizer denormalizer used to transform array to object
+     * @param ObjectDetacherInterface            $detacher     detacher to remove it from UOW when skip
+     * @param string                             $class        class of the object to instanciate in case if need
+     * @param string                             $format       format use to denormalize
+     */
+    public function __construct(
+        ReferableEntityRepositoryInterface $repository,
+        DenormalizerInterface $denormalizer,
+        ValidatorInterface $validator,
+        ObjectDetacherInterface $detacher,
+        $class, // TODO responsibility of the denormalizer ?!
+        $format
+    ) {
+        parent::__construct($repository, $denormalizer, $validator, $detacher, $class);
+        $this->format = $format;
+    }
 
     /**
      * {@inheritdoc}
@@ -86,7 +110,7 @@ class GroupProcessor extends AbstractProcessor
         $group = $this->denormalizer->denormalize(
             $groupData,
             $this->class,
-            'csv',
+            $this->format,
             ['entity' => $group]
         );
