+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/processors.yml
@@ -6,8 +6,8 @@ parameters:
     pim_base_connector.processor.transformer.class:                   Pim\Bundle\BaseConnectorBundle\Processor\TransformerProcessor
     pim_base_connector.processor.product.class:                       Pim\Bundle\BaseConnectorBundle\Processor\ProductProcessor
     pim_base_connector.processor.category.class:                      Pim\Bundle\BaseConnectorBundle\Processor\CategoryProcessor
-    pim_base_connector.processor.array_to_object.flat.group.class:         Pim\Bundle\BaseConnectorBundle\Processor\ArrayToObject\Flat\GroupProcessor
-    pim_base_connector.processor.array_to_object.flat.variant_group.class: Pim\Bundle\BaseConnectorBundle\Processor\ArrayToObject\Flat\VariantGroupProcessor
+    pim_base_connector.processor.denormalization.flat.group.class:         Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\GroupProcessor
+    pim_base_connector.processor.denormalization.flat.variant_group.class: Pim\Bundle\BaseConnectorBundle\Processor\Denormalization\VariantGroupProcessor
 
     pim_base_connector.processor.product_to_flat_array.class:        Pim\Bundle\BaseConnectorBundle\Processor\ProductToFlatArrayProcessor
     pim_base_connector.processor.dummy.class:                        Pim\Bundle\BaseConnectorBundle\Processor\DummyProcessor
@@ -120,23 +120,25 @@ services:
         calls:
             - [ addMapping, [ configuration, rawConfiguration ] ]
 
-    # deprecated, will be dropped in 1.4, use pim_base_connector.processor.array_to_object.flat.group or variant_group
+    # deprecated, will be dropped in 1.4, use pim_base_connector.processor.denormalization.flat.group or variant_group
     pim_base_connector.processor.group:
         parent: pim_base_connector.processor.transformer
         arguments:
             - %pim_catalog.entity.group.class%
 
-    pim_base_connector.processor.array_to_object.flat.group:
-        class: %pim_base_connector.processor.array_to_object.flat.group.class%
+    # TODO name flat or csv ?
+    pim_base_connector.processor.denormalization.flat.group:
+        class: %pim_base_connector.processor.denormalization.flat.group.class%
         arguments:
             - '@pim_catalog.repository.group'
             - '@pim_serializer'
             - '@pim_validator'
             - '@akeneo_storage_utils.doctrine.object_detacher'
             - %pim_catalog.entity.group.class%
+            - 'csv'
 
-    pim_base_connector.processor.array_to_object.flat.variant_group:
-        class: %pim_base_connector.processor.array_to_object.flat.variant_group.class%
+    pim_base_connector.processor.denormalization.flat.variant_group:
+        class: %pim_base_connector.processor.denormalization.flat.variant_group.class%
         arguments:
             - '@pim_catalog.repository.group'
             - '@pim_serializer'
@@ -145,3 +147,4 @@ services:
             - '@akeneo_storage_utils.doctrine.object_detacher'
             - %pim_catalog.entity.group.class%
             - %pim_catalog.entity.product_template.class%
+            - 'csv'
